Models creating

+++ scala-http4s/src/main/scala/su/wps/blog/migrations/Migrate_20180915142027_CreateCommentRaters.scala
@@ -7,7 +7,7 @@ class Migrate_20180915142027_CreateCommentRaters extends Migration {
 
   def up(): Unit = {
     createTable(tableName) { t =>
-      t.integer("id", PrimaryKey, AutoIncrement)
+      t.bigint("id", PrimaryKey, AutoIncrement)
       t.varchar("ip", Limit(39), NotNull)
       t.integer("comment_id", NotNull)
     }
